requirements.txt
streamlit
google-generativeai
groq
pymupdf
langchain
langchain-community
sentence-transformers
chromadb
python-dotenv
